Add ProductPlaceholder component with image and description layout
